+++ plugins/contentbank/classes/form.php
@@ -230,6 +230,25 @@ class form extends \mod_interactivevideo\form\base_form {
             'hascompletion' => true,
         ]);
 
+        // Save state.
+        $groups = [];
+        $groups[] = $mform->createElement(
+            'advcheckbox',
+            'savecurrentstate',
+            '',
+            get_string('yes'),
+            ['group' => 1],
+            [0, 1]
+        );
+        $groups[] = $mform->createElement(
+            'static',
+            'savecurrentstatedesc',
+            '',
+            '<span class="text-muted small w-100 d-block">'
+                . get_string('savecurrentstatedesc', 'ivplugin_contentbank') . '</span>'
+        );
+        $mform->addGroup($groups, '', get_string('savecurrentstate', 'ivplugin_contentbank'), null, false);
+
         $elements = [];
         $elements[] = $mform->createElement(
             'advcheckbox',
@@ -384,6 +403,19 @@ class form extends \mod_interactivevideo\form\base_form {
         $mform->hideIf('textonfailed', 'showtextonfailed', 'eq', 0);
 
         $this->close_form();
+    }
+
+    /**
+     * Process advanced settings
+     *
+     * @param \stdClass $data
+     * @return string
+     */
+    public function process_advanced_settings($data) {
+        $adv = parent::process_advanced_settings($data);
+        $adv = json_decode($adv);
+        $adv->savecurrentstate = $data->savecurrentstate;
+        return json_encode($adv);
     }
 
     /**